Added snap functionality
Added a "cursorPreprocessor" option where you can pass a function to modify the cursor event before it gets used in the draw functions. This allows for us to use a function to snap to lines or points.

Also modified the rendering to show start and intermediate points (so it's clear when you're snapping or not.)